social_backend/pyproject.toml
[tool.poetry]
name = "social_backend"
version = "0.1.0"
description = ""
authors = ["Dmitriy Mushtruk <dima.mushtruk@gmail.com>"]
readme = "README.md"
packages = [{include = "social_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.3"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
django-cors-headers = "4.1.0"
psycopg2-binary = "2.9.6"
pillow = "^10.0.0"
pylint = "^2.17.5"
pylint-django = "^2.5.3"
ptvsd = "^4.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
